+++ src/mlpack/methods/cf/normalization/user_mean_normalization.hpp
@@ -112,9 +112,9 @@ class UserMeanNormalization
       double tmp = *it - userMean(it.col());
 
       // The algorithm omits rating of zero. If normalized rating equals zero,
-      // it is set to the smallest positive double value.
+      // it is set to the smallest positive float value.
       if (tmp == 0)
-        tmp = std::numeric_limits<double>::min();
+        tmp = std::numeric_limits<float>::min();
 
       *it = tmp;
     }
